+++ services/src/api/handlers/permissions.rs
@@ -2,9 +2,9 @@ use crate::api::model::datatypes::{DatasetId, LayerId};
 use crate::contexts::{ApplicationContext, SessionContext};
 use crate::error::Result;
 use crate::layers::listing::LayerCollectionId;
+use crate::permissions::{Permission, PermissionListing};
+use crate::permissions::{PermissionDb, ResourceId, RoleId};
 use crate::pro::contexts::{ProApplicationContext, ProGeoEngineDb};
-use crate::pro::permissions::{Permission, PermissionListing};
-use crate::pro::permissions::{PermissionDb, ResourceId, RoleId};
 use crate::projects::ProjectId;
 use actix_web::{web, FromRequest, HttpResponse};
 use geoengine_datatypes::error::BoxedResultExt;
