+++ services/src/pro/mod.rs
@@ -3,7 +3,6 @@
 pub mod contexts;
 pub mod datasets;
 pub mod layers;
-pub mod permissions;
 pub mod projects;
 pub mod users;
 pub mod util;
