+++ services/src/pro/users/session.rs
@@ -1,8 +1,8 @@
 use crate::api::handlers::get_token;
 use crate::contexts::{ApplicationContext, Session, SessionId};
 use crate::error;
+use crate::permissions::{Role, RoleId};
 use crate::pro::contexts::PostgresContext;
-use crate::pro::permissions::{Role, RoleId};
 use crate::pro::users::UserId;
 use crate::projects::{ProjectId, STRectangle};
 use crate::util::Identifier;
